+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed 
 - index out of range in the Config.py migration code #10
 - LinuxCNC expect to see % as the first character and the last of the program #11
+- MirrorGcode.py:43: SyntaxWarning: invalid escape sequence '\-' #13
 
 ## [0.1.9] - 2025-07-28